third_party/blink/web_tests/fast/css/invalidation/unresolved-pseudo-expected.txt

CONSOLE WARNING: line 10: The Custom Elements v0 API (:unresolved pseudo selector) has been removed. See https://www.chromestatus.com/feature/4642138092470272 for more details.
Use invalidation sets for :unresolved pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(child).backgroundColor is red
PASS getComputedStyle(sibling).backgroundColor is red
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(child).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(sibling).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
